+++ ExpandWorldSize/ExpandWorldSize.cs
@@ -12,7 +12,7 @@ public class EWS : BaseUnityPlugin
 {
   public const string GUID = "expand_world_size";
   public const string NAME = "Expand World Size";
-  public const string VERSION = "1.7";
+  public const string VERSION = "1.8";
 #nullable disable
   public static ManualLogSource Log;
   public static Harmony harmony;
@@ -34,6 +34,7 @@ public class EWS : BaseUnityPlugin
     if (NeedsMigration)
       MigrateOldConfig();
     harmony = new(GUID);
+    harmony.Patch(AccessTools.Method(typeof(Terminal), nameof(Terminal.InitTerminal)), null, new HarmonyMethod(typeof(DebugCommands), nameof(DebugCommands.Add)));
     try
     {
       SetupWatcher();