+++ ExpandWorldSize/config/ConfigWrapper.cs
@@ -18,7 +18,7 @@ public class ConfigWrapper
     ConfigSync = configSync;
     Regenerate = regenerate;
   }
-  public static Dictionary<ConfigEntry<string>, float> Floats = new();
+  public static Dictionary<ConfigEntry<string>, float?> Floats = new();
   public static Dictionary<ConfigEntry<string>, int?> Ints = new();
 
   public ConfigEntry<string> BindFloat(string group, string name, float value, bool regenerate, string description = "", bool synchronizedSetting = true)
@@ -28,11 +28,11 @@ public class ConfigWrapper
     Floats[entry] = TryParseFloat(entry);
     return entry;
   }
-  public ConfigEntry<string> BindInt(string group, string name, int? value, bool regenerate, string description = "", bool synchronizedSetting = true)
+  public ConfigEntry<string> BindFloat(string group, string name, float? value, bool regenerate, string description = "", bool synchronizedSetting = true)
   {
-    var entry = Bind(group, name, value.HasValue ? value.ToString() : "", regenerate, description, synchronizedSetting);
-    entry.SettingChanged += (s, e) => Ints[entry] = TryParseInt(entry);
-    Ints[entry] = TryParseInt(entry);
+    var entry = Bind(group, name, value.HasValue ? value.Value.ToString(CultureInfo.InvariantCulture) : "", regenerate, description, synchronizedSetting);
+    entry.SettingChanged += (s, e) => Floats[entry] = TryParseFloat(entry);
+    Floats[entry] = TryParseFloat(entry);
     return entry;
   }
 
@@ -48,20 +48,14 @@ public class ConfigWrapper
   public ConfigEntry<T> Bind<T>(string group, string name, T value, bool regenerate, string description = "", bool synchronizedSetting = true) => Bind(group, name, value, regenerate, new ConfigDescription(description), synchronizedSetting);
   private void Regen(object e, EventArgs s) => Regenerate();
 
-
-  private static int? TryParseInt(ConfigEntry<string> setting)
-  {
-    if (int.TryParse(setting.Value, NumberStyles.Integer, CultureInfo.InvariantCulture, out var result)) return result;
-    return null;
-  }
-  private static float TryParseFloat(string value, float defaultValue)
+  private static float? TryParseFloat(string value)
   {
     if (float.TryParse(value, NumberStyles.Float, CultureInfo.InvariantCulture, out var result)) return result;
-    return defaultValue;
+    return null;
   }
-  private static float TryParseFloat(ConfigEntry<string> setting)
+  private static float? TryParseFloat(ConfigEntry<string> setting)
   {
     if (float.TryParse(setting.Value, NumberStyles.Float, CultureInfo.InvariantCulture, out var result)) return result;
-    return TryParseFloat((string)setting.DefaultValue, 0f);
+    return TryParseFloat((string)setting.DefaultValue);
   }
 }